# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JamberTech WhatsApp Session Generator API
 * OpenAPI spec version: 2.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface PairRequest {
  /** Phone number with country code (e.g. 923001234567) */
  number: string;
}

export interface PairResponse {
  pairCode: string;
  sessionId: string;
}

export interface QrResponse {
  sessionId: string;
  /** Base64 data URL of QR code image */
  qr: string;
}

export type SessionStatusResponseStatus =
  (typeof SessionStatusResponseStatus)[keyof typeof SessionStatusResponseStatus];

export const SessionStatusResponseStatus = {
  waiting: "waiting",
  connected: "connected",
  expired: "expired",
} as const;

export interface SessionStatusResponse {
  status: SessionStatusResponseStatus;
  sessionID?: string;
  qr?: string;
}

export interface ErrorResponse {
  error: string;
}