# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 2.0.0
  description: JamberTech WhatsApp Session Generator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: whatsapp
    description: WhatsApp session operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /pair:
    post:
      operationId: generatePairCode
      tags: [whatsapp]
      summary: Generate WhatsApp pair code
      description: Generates a pair code for linking WhatsApp via phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairRequest"
      responses:
        "200":
          description: Pair code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /qr:
    post:
      operationId: generateQr
      tags: [whatsapp]
      summary: Generate WhatsApp QR code
      description: Generates a QR code for linking WhatsApp
      responses:
        "200":
          description: QR code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /qr/{token}:
    get:
      operationId: pollQrStatus
      tags: [whatsapp]
      summary: Poll QR connection status
      description: Poll the status of a QR code session
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: QR status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStatusResponse"
  /status/{token}:
    get:
      operationId: pollPairStatus
      tags: [whatsapp]
      summary: Poll pair code connection status
      description: Poll the status of a pair code session
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pair status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStatusResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    PairRequest:
      type: object
      properties:
        number:
          type: string
          description: Phone number with country code (e.g. 923001234567)
      required:
        - number
    PairResponse:
      type: object
      properties:
        pairCode:
          type: string
        sessionId:
          type: string
      required:
        - pairCode
        - sessionId
    QrResponse:
      type: object
      properties:
        sessionId:
          type: string
        qr:
          type: string
          description: Base64 data URL of QR code image
      required:
        - sessionId
        - qr
    SessionStatusResponse:
      type: object
      properties:
        status:
          type: string
          enum: [waiting, connected, expired]
        sessionID:
          type: string
        qr:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error