# lib/api-zod/src/generated/types/pairResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JamberTech WhatsApp Session Generator API
 * OpenAPI spec version: 2.0.0
 */

export interface PairResponse {
  pairCode: string;
  sessionId: string;
}