# lib/api-zod/src/generated/types/qrResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JamberTech WhatsApp Session Generator API
 * OpenAPI spec version: 2.0.0
 */

export interface QrResponse {
  sessionId: string;
  /** Base64 data URL of QR code image */
  qr: string;
}