# lib/api-zod/src/generated/types/sessionStatusResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JamberTech WhatsApp Session Generator API
 * OpenAPI spec version: 2.0.0
 */
import type { SessionStatusResponseStatus } from "./sessionStatusResponseStatus";

export interface SessionStatusResponse {
  status: SessionStatusResponseStatus;
  sessionID?: string;
  qr?: string;
}